+++ src/irvine/oeis/a292/A292280.java
@@ -23,6 +23,6 @@ public class A292280 extends GeneralizedEulerTransform {
 
   @Override
   protected Z advanceG(final long k) {
-    return Functions.FACTORIAL.z((int) k);
+    return Functions.FACTORIAL.z(k);
   }
 }
